fastapi/pyproject.toml
[project]
name = "fastapi-labs"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "Ittichai Boonyarakthunya",email = "i.boonyarakthunya@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "create-fastapi-project (>=0.2.5,<0.3.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "prisma (>=0.15.0,<0.16.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "redis (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.0"
pre-commit = "^4.2.0"

